// src/app/DTOs/word-data.DTO.ts
export interface WordData {
  word: string;
  phonetic: string;
  phonetics: [
    {
      text: string;
      audio: string;
    },
    {
      text: string;
    }
  ];
  origin: string;
  meanings: [
    {
      partOfSpeech: string;
      definitions: [
        {
          definition: string;
          example: string;
          synonyms: [];
          antonyms: [];
        }
      ];
    },
    {
      partOfSpeech: string;
      definitions: [
        {
          definition: string;
          example: string;
          synonyms: [];
          antonyms: [];
        }
      ];
    },
    {
      partOfSpeech: string;
      definitions: [
        {
          definition: string;
          example: string;
          synonyms: [];
          antonyms: [];
        }
      ];
    }
  ];
}